fix(implement-apollo-server): Fix Apollo Server - Remove buildSchema and introduce typeDefs - Pass models through context
