+++ next/contributors-site/static/contributors.json
@@ -13,6 +13,10 @@
       "name": "Temilayo Subair",
       "twitter":"https://twitter.com/temi_subair",
       "github":"https://github.com/fersubair40"
+    },
+    {
+      "name": "Benny Ogidan",
+      "twitter": "https://twitter.com/BennyOgidan",
+      "github": "https://github.com/benfluleck"
     }
   ]
-
