+++ node/greek-mythology/appwithgraphqlusingapollo/models/index.js
@@ -1,0 +1,5 @@
+const GreekGod = require('./greekgod');
+
+module.exports = {
+  GreekGod
+}
